Create ForumThread.cshtml
This file is a webpage that I worked on with my friend Steven Evenhuis. The code is for a particular forum thread that would show all posts under the thread. The HTML was written mostly by Steven, where as I am responsible for a lot of the JavaScript, specifically the functions that involve manipulating the UI, as well as those that handle data submission and callbacks. Using the combination of SignalR and AngularJS, we were able to make a very slick UI with some interesting features for the user, such as real time appending of new posts as well as visual feedback for upvotes and downvotes. Additionally, I created a very nice user experience by having multiple ways to open and close the reply box.